Fix quantity

+++ migrations/mysql/modelstg/20220412191255376_create_table_units_distribution.sql
@@ -4,7 +4,7 @@ CREATE TABLE `units_distribution` (
     `date` datetime NOT NULL,
     `logistic_center_id` varchar(45) NOT NULL,
     `process_name` varchar(45) NOT NULL,
-    `quantity` bigint(8) NOT NULL,
+    `quantity` double(8,3) NOT NULL,
     `quantity_metric_unit` varchar(45) NOT NULL,
      PRIMARY KEY (`id`)
 );
